requirements.txt
torch
opencv-python
ultralytics
pillow

